conference proposal
git-svn-id: svn+ssh://yazoonweb/home/svn/projects/iws@508 73cf1824-746b-4eb0-ba16-0c255e03c809

+++ src/huard/iws/bean/ConferenceProposalBean.java
@@ -52,9 +52,9 @@ public class ConferenceProposalBean {
 	private int initiatingBodyRole;
 	private long openDate;
 	private long submissionDate;
-	private double totalCost;
+	private int totalCost;
 	private int totalCostCurrency;
-	private double supportSum;
+	private int supportSum;
 	private int supportCurrency;
 	private boolean auditorium;
 	private boolean seminarRoom;
@@ -632,19 +632,19 @@ public class ConferenceProposalBean {
 		this.toDate = toDate;
 	}
 
-	public double getTotalCost() {
+	public int getTotalCost() {
 		return totalCost;
 	}
 
-	public void setTotalCost(double totalCost) {
+	public void setTotalCost(int totalCost) {
 		this.totalCost = totalCost;
 	}
 
-	public double getSupportSum() {
+	public int getSupportSum() {
 		return supportSum;
 	}
 
-	public void setSupportSum(double supportSum) {
+	public void setSupportSum(int supportSum) {
 		this.supportSum = supportSum;
 	}
